docs: update student features with modular exam hooks
