+++ README.md
@@ -83,11 +83,12 @@ docker-compose -p psychology-test-platform up -d
 - ✅ 考试管理：发布、密码保护、时间限制
 - ✅ 结果分析：统计查看、数据导出
 
-### 👨‍🎓 学生端功能  
+### 👨‍🎓 学生端功能
 - ✅ 无需注册：公共链接直接答题
 - ✅ 条件显示：根据答案动态显示题目
 - ✅ 进度保存：本地存储防止数据丢失
 - ✅ 防重复提交：同一学生只能提交一次
+- ✅ 组件化考试流程：拆分 `useExamFlow` 与 `useAIWebRTC` 等独立 Hook
 
 ## 🛠️ 开发指南
