+++ frontend/src/components/ParticipantExam/index.ts
@@ -3,7 +3,14 @@ export { default as ExamStateManager } from './ExamStateManager';
 export { default as QuestionNavigation } from './QuestionNavigation';
 export { default as ExamSubmissionManager } from './ExamSubmissionManager';
 export { default as QuestionRenderer } from './QuestionRenderer';
+export { default as AIStatusPanel } from './AIStatusPanel';
+
+// Hook 导出
+export { useExamFlow } from './useExamFlow';
+export { useAIWebRTC } from './useAIWebRTC';
 
 // 类型导出
 export type { ExamInfo, ParticipantInfo, ExamStep } from './ExamStateManager';
 export type { ExamSubmissionManagerProps } from './ExamSubmissionManager';
+export type { UseExamFlowReturn } from './useExamFlow';
+export type { UseAIWebRTCReturn } from './useAIWebRTC';
